feat: remove `AddressParseError` and add private `anonymize_str()`
- Removed `AddressParseError` variant altogether
  - `CryptoPAnError` now has only 1 variants, `CipherError`
  - needs to be either "flattened" or refactored later

`CryptoPAn`'s job is to anonymize IP addresses, and having an IP string
parsing error in its error variants sounds awkward and not DRY.

- Added `anonymize_str(&str)` as a **private method**
  - Unit tests now makes use of the said method

As a private because we need to decide if it's really necessary.
Panicking convenience method in library API is considered unidiomatic.

`chrono` APIs used to panic and a separate `_opt()` variants were used
to return errors instead, such as TimeZone::ymd()` and `ymd_opt()`.
But later they were all deprecated and current APIs all return errors.

If your clients really insists on having ergonomics, welp, `pub` it is.
Otherwise this should be moved down to the test module as a helper func.